Post transaction summary is logged for API users (RhBug:1855158)
Post transaction summary is always logged into /var/log/dnf.log.
When transaction is called from cli, the summary is also printed to
stdout in columns (as previously).

= changelog =
msg:           Packages installed/removed via DNF API are logged into dnf.log
type:          enhancement
resolves:      https://bugzilla.redhat.com/show_bug.cgi?id=1855158
